feat(app/server): add production app initialization module
Introduced a dedicated `production-app` module to streamline
production app setup. This refactor replaces inline handling in
`server.ts`, improving modularity and maintainability. Also removed
redundant methods and simplified environment-based app initialization.
